refactor: move folder to dtos
